Revert "Merge pull request #193 from codemod-js/semantic-release-monorepo"
This reverts commit b58704cc00e099ed320aabda197be292a4e30697, reversing
changes made to 23839c9829ec959dce32501f701ebe7f7c93e096.
