+++ script/_utils/runInPackages.ts
@@ -1,7 +1,7 @@
 import * as globby from 'globby';
+import { spawn } from 'child_process';
 import { readFile } from 'mz/fs';
 import { join } from 'path';
-import runInDirectory from './runInDirectory';
 
 export default async function runInPackages(
   stdin: NodeJS.ReadStream,
@@ -12,8 +12,41 @@ export default async function runInPackages(
   packages?: Array<string>
 ): Promise<void> {
   for (const pkg of await (packages || findPackages())) {
-    await runInDirectory(stdin, stdout, stderr, pkg, command, args);
+    await runForPackage(stdin, stdout, stderr, pkg, command, args);
   }
+}
+
+async function runForPackage(
+  stdin: NodeJS.ReadStream,
+  stdout: NodeJS.WriteStream,
+  stderr: NodeJS.WriteStream,
+  pkg: string,
+  command: string,
+  args: Array<string> = []
+): Promise<void> {
+  return new Promise<void>((resolve, reject) => {
+    stdout.write(`${pkg} ❯ ${command} ${args.join(' ')}\n`);
+
+    const child = spawn(
+      command,
+      args,
+      {
+        cwd: pkg,
+        stdio: [stdin, stdout, stderr]
+      });
+
+    child.on('exit', code => {
+      if (code === 0) {
+        resolve();
+      } else {
+        reject(new Error(`exit status was ${code} for command: ${command} args: [${args.join(' ')}]`));
+      }
+    });
+
+    child.on('error', error => {
+      reject(error);
+    });
+  });
 }
 
 async function findPackages(): Promise<Array<string>> {
